test: split isPostalCode validator tests
